requirements.txt
# Core libraries
numpy==1.24.3
pandas==1.5.3

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2

# Text preprocessing and NLP
nltk==3.8.1
scikit-learn==1.2.2
gensim==4.3.1

# Recommendation system utilities
scipy==1.10.1
tensorflow==2.13.0
keras==2.13.1

# Web scraping (if needed for additional data)
beautifulsoup4==4.12.2
requests==2.31.0

# Web app development
streamlit==1.25.0

# Utilities
pyyaml==6.0
tqdm==4.66.0

# Logging and monitoring
loguru==0.7.0

# Testing
pytest==7.4.0

# Database support
sqlalchemy==2.0.20
sqlite-utils==3.33

# Deployment (optional, for deploying via Docker or Heroku)
gunicorn==21.2.0

# For Kaggle dataset ingestion
kaggle==1.5.15
kagglehub

# Optional visualization tools
plotly==5.15.0

-e .